% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.measure.R
\name{discretize2d}
\alias{discretize2d}
\title{Discretize 2-dimensional continuous data into bins}
\usage{
discretize2d(x, y, model = c("uniform_width", "uniform_frequency"))
}
\arguments{
\item{x}{a numeric vector of the random variable x.}

\item{y}{a numeric vector of the random variable y.}

\item{model}{two discretization algorithms are available, "uniform_width" is the default.}
}
\value{
discretize2d returns a 2-dimensional count table.
}
\description{
The function of discretize2d is used to assign the observations of two sets of continuous random variables to bins,
and returns a corresponding two-dimensional count table. Two of the most common discretization methods are available:
"uniform width" and "uniform frequency".
}
\details{
Uniform width-based method ("uniform_width") divides the continuous data into N bins with equal width.
Uniform frequency-based method ("uniform_frequency") divides the continuous data into N bins with (approximate) equal count number.
By default in both methods, the number of bins N is initialized into a round-off value according to the square root of the data size.
}
\examples{

# load Informeasure library
library("Informeasure")

# two numeric vectors that correspond to two continuous random variables
x <- c(0.0, 0.2, 0.2, 0.7, 0.9, 0.9, 0.9, 0.9, 1.0)
y <- c(1.0, 2.0,  12, 8.0, 1.0, 9.0, 0.0, 3.0, 9.0)

# corresponding count table estimated by "uniform width" algorithm
discretize2d(x,y, "uniform_width")

# corresponding count table estimated by "uniform frequency" algorithm
discretize2d(x,y, "uniform_frequency")
}
